Move reuse macro to lib.rs tentatively.
This is because doing "macro_export" make those macros visible from
outside of the crate, and this is not good in terms of future-proof.
